+++ APP/CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.3
+
+Added instructions to sidebar action.
+Added 'Reset' and 'Skip All' buttons.
+Added visual feedback to list items and buttons.
+
 ## v0.2
 
 Added sidebar action for marking tabs to skip during reload.
@@ -6,3 +12,4 @@ Added context menu to easily access sidebar.
 ## v0.1
 
 Simple browser action with icon.
+1